% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_ids.R
\name{add_auto_columns}
\alias{add_auto_columns}
\alias{hash_id}
\title{Add columns with auto-generated ids}
\usage{
add_auto_columns(table, table_name, model, nchar = 8)

hash_id(x, nchar = 8)
}
\arguments{
\item{table}{Data table to import (tibble or data.frame)}

\item{table_name}{Name of data table in model}

\item{model}{\code{\link{dm}} object describing data model}

\item{nchar}{number of characters in the resulting strings (max 32)}

\item{x}{string use to create hash}
}
\value{
\code{table} with additional columns

identifier based on a hash of \code{x}
}
\description{
Add columns to a table using a hash of other columns, as defined in a data model
}
\details{
Automatically generated hashed identifiers based on the values of
other columns in \code{table}. \code{model} should have an attribute
\code{auto_id} containing a vector of columns to be used.
}
