% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_data_tables.R
\name{edit_data_tables}
\alias{edit_data_tables}
\alias{delete_rows}
\title{Edit data tables}
\usage{
delete_rows(
  pk_value,
  table_name,
  pk_name = paste0(table_name, "_id"),
  tables,
  model
)
}
\arguments{
\item{pk_value}{Value of the primary key to delete}

\item{table_name}{Name of the table to delete from}

\item{pk_name}{Name of the primary key column. Defaults to the table name + "_id".}

\item{tables}{Named list of data tables}

\item{model}{\code{\link{dm}} object describing data model}
}
\value{
\code{\link{dm}} object with filtered tables
}
\description{
Edit a set of data tables while maintaining consistency among foreign keys
}
\details{
When calling \code{delete_rows}, a side effect is that only rows that are
related via foreign keys to rows in \code{table_name} are kept. Therefore,
\code{table_name} should always be the primary table.
}
