% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_model.R
\name{hash_id}
\alias{hash_id}
\title{create a hashed identifier}
\usage{
hash_id(x, nchar = 8)
}
\arguments{
\item{x}{string use to create hash}

\item{nchar}{number of characters in the resulting string}
}
\value{
identifier based on a hash of \code{x}
}
\description{
create a hashed identifier
}
