% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_model.R
\name{transpose_field_value}
\alias{transpose_field_value}
\title{transpose field,value pairs to a data model}
\usage{
transpose_field_value(fv, table_name, model)
}
\arguments{
\item{fv}{data frame with columns field, value}

\item{table_name}{name of table in model}

\item{model}{\code{\link{dm}} object describing data model}
}
\value{
single row data frame with names from \code{fv$field} and values from \code{fv$value}
}
\description{
transpose field,value pairs to a data model
}
