% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anvil_import.R
\name{anvil_import}
\alias{anvil_import}
\alias{anvil_import_table}
\alias{anvil_import_set}
\alias{create_set_all}
\alias{unnest_set_table}
\alias{anvil_import_tables}
\title{Import data tables into AnVIL}
\usage{
anvil_import_table(
  table,
  table_name,
  model,
  overwrite = FALSE,
  namespace = avworkspace_namespace(),
  name = avworkspace_name()
)

anvil_import_set(
  table,
  table_name,
  overwrite = FALSE,
  namespace = avworkspace_namespace(),
  name = avworkspace_name()
)

create_set_all(table, table_name)

unnest_set_table(set_table)

anvil_import_tables(
  tables,
  model,
  overwrite = FALSE,
  namespace = avworkspace_namespace(),
  name = avworkspace_name()
)
}
\arguments{
\item{table}{Data table to import (tibble or data.frame)}

\item{table_name}{Name of data table in model}

\item{model}{\code{\link{dm}} object describing data model}

\item{overwrite}{Logical for whether to overwrite data for existing rows}

\item{namespace}{AnVIL workspace namespace}

\item{name}{AnVIL workspace name}

\item{set_table}{set table returned by \code{\link{avtable}}}

\item{tables}{Named list of tables to import}
}
\value{
\code{create_set_all} returns a tibble with one set, 'all',
containing all entities in \code{table}.

\code{unnest_set_table} returns a tibble with a set table in
long form
}
\description{
Import data tables into AnVIL
}
\details{
\code{anvil_import_table} imports a data table to an AnVIL workspace.

\code{anvil_import_set} imports a set table to an AnVIL workspace. The
reference table must already exist in the workspace.

\code{anvil_import_tables} imports a named list of tables, calling \code{anvil_import_table}
or \code{anvil_import_set} as appropriate. Set tables are assumed to end with "_set".

\code{create_set_all} creates a set table containing all entities in the
reference table.

\code{unnest_set_table} takes a set table returned by \code{\link{avtable}}
and formats it so it can be imported back into AnVIL with 
\code{\link{avtable_import_set}}
}
