% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_model.R
\name{import_tsv}
\alias{import_tsv}
\alias{tsv_to_dm}
\alias{tsv_to_dbml}
\title{Import data from tsv file}
\usage{
tsv_to_dm(tsv)

tsv_to_dbml(tsv, dbml)
}
\arguments{
\item{tsv}{Character vector with paths to tab-separated variable files. If more than one file is provided, they will be combined into a single data model.}

\item{dbml}{Path for the DBML output file}
}
\value{
\code{\link{dm}} object
}
\description{
Import data model in tsv format
}
\details{
\code{tsv_to_dm} returns a \code{\link{dm}} object.

\code{tsv_to_dbml} writes a \href{https://www.dbml.org}{DBML} file.

TSV files to be imported must have the following columns: 
\itemize{
  \item{entity: }{"Table", "enum", or "meta"}
  \item{table: }{table or enum name}
  \item{column: }{column name within table}
  \item{type: }{"string", "boolean, "integer", "float", "date", "dateTime", or name of enum}
  \item{required: }{TRUE indicates the column is required, FALSE or missing if the column
   is optional. 'CONDITIONAL (column = value)' indicates a requirement only if any element of 
   'column' contains 'value'. To indicate that table 't2' is only required if table 't1' is 
   present, entity='meta', table='t2' and required='CONDITIONAL (t1)'}
  \item{pk: }{logical where TRUE indicates the column is a primary key, 
    other values may be FALSE or missing}
  \item{ref: }{Reference to other columns in the data model. Either 
    1) a cross-table reference string following the 
    \href{https://www.dbml.org/docs/#relationships-foreign-key-definitions}{DBML}
    format, or 
    2) 'from:' followed by a comma-separated list of columns used to
    automatically generate this column using a hash function.}
  \item{note}{Note string (column description) following the
    \href{https://www.dbml.org/docs/#column-notes}{DBML} format}
}
}
\examples{
tsv <- system.file("extdata", "data_model.tsv", package="AnvilDataModels")
(dm <- tsv_to_dm(tsv))
attr(dm, "required")
lapply(dm, attr, "required")

tmp <- tempfile()
tsv_to_dbml(tsv, tmp)
readLines(tmp, n=14)

unlink(tmp)

tsv <- system.file("extdata", "data_model_conditional.tsv", package="AnvilDataModels")
(dm <- tsv_to_dm(tsv))
attr(dm, "conditions")
lapply(dm, attr, "conditions")

}
